app/requirements.txt
Django==2.1
gunicorn
mongoengine
djongo
redis==3.2.1
huey==2.0.1
natsort
Pillow
pypdf2
google-cloud-vision
nltk
matplotlib
beautifulsoup4
djangorestframework
django-rest-framework-mongoengine
pymysql
neo4j
elasticsearch
lxml
